feat(#47): post 라우터 이름 변경
